move om_lines namespace to overmap.h
